+++ tests/test_testing_models.py
@@ -86,7 +86,7 @@ def test_calculate_positive_tests_to_distribute_per_age_group():
         test_shares_by_age_group,
         positivity_rate_by_age_group,
     )
-    expected = pd.Series([1, 4, 5], index=["0-4", "5-14", "15-34"])
+    expected = pd.Series([1, 4, 5], index=["0-4", "5-14", "15-34"], dtype="int32")
     pd.testing.assert_series_equal(res, expected, check_names=False)
